Represent numbers as floats for now

+++ core/src/builtins/number.rs
@@ -1,18 +1,17 @@
 use crate::*;
-use bigdecimal::BigDecimal;
 
 #[derive(TypeInfo, Debug, Clone)]
 pub struct Number {
-    pub number: BigDecimal,
+    pub number: f64,
     pub location: Option<SourceLocation>,
 }
 
 impl Number {
-    pub fn new(number: BigDecimal) -> Self {
+    pub fn new(number: f64) -> Self {
         Number::new_located(number, None)
     }
 
-    pub fn new_located(number: BigDecimal, location: Option<SourceLocation>) -> Self {
+    pub fn new_located(number: f64, location: Option<SourceLocation>) -> Self {
         Number { number, location }
     }
 }
